lft_classify/requirements.txt
keras_cv
matplotlib
pillow
